fix : 가격과 수량 Long에서 Integer로 변경

+++ src/main/java/com/postick/kiosk/domain/Order.java
@@ -19,5 +19,5 @@ public class Order {
 	private Long id;
 
 	@Column(nullable = false)
-	private Long totalPrice;
+	private Integer totalPrice;
 }
